espnet-asr: add Python 3.8 compatibility
The type hinting syntax is available since Python 3.9. For this
reason, it ended up causing an execption on Python 3.8.

    TypeError: 'type' object is not subscriptable

Fix it thusly. Tested on Ubuntu 20.04.

Note: This patch is only for ESPnet module. Since the oldest Python
version nvidia/nemo supports is 3.10, pkg/nemo-asr does not support
the older Python versions.

Signed-off-by: Fujimoto Seiji <fujimoto@ceptord.net>